requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==1.1.3.post0
idna==3.4
infi.clickhouse-orm==2.1.2
iso8601==1.1.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
SQLAlchemy==1.4.42
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
mypy==0.982
mypy-extensions==0.4.3
